Fixes namespaces and Class names
