--- FoodFite/EmptyBot.cs
@@ -1,60 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-//
-// Generated with EmptyBot .NET Template version v4.10.3
-
-using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Bot.Builder;
-using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Schema;
-
-namespace EmptyBot
-{
-    public class EmptyBot<T> : ActivityHandler where T: Dialog
-    {
-        protected readonly Dialog Dialog;
-        protected readonly BotState ConversationState;
-        protected readonly BotState UserState;
-
-        public EmptyBot(ConversationState conversationState, UserState userState, T dialog)
-        {
-            ConversationState = conversationState;
-            UserState = userState;
-            Dialog = dialog;
-        }
-
-        public override async Task OnTurnAsync(ITurnContext turnContext, CancellationToken cancellationToken = default)
-        {
-            await base.OnTurnAsync(turnContext, cancellationToken);
-
-            // Save any state changes that might have occurred during the turn.
-            await ConversationState.SaveChangesAsync(turnContext, false, cancellationToken);
-            await UserState.SaveChangesAsync(turnContext, false, cancellationToken);
-        }
-
-        // protected override async Task OnMembersAddedAsync(IList<ChannelAccount> membersAdded, ITurnContext<IConversationUpdateActivity> turnContext, CancellationToken cancellationToken)
-        // {
-           
-        //     foreach (var member in membersAdded)
-        //     {
-        //         if (member.Id != turnContext.Activity.Recipient.Id)
-        //         {
-        //             await Dialog.RunAsync(turnContext, ConversationState.CreateProperty<DialogState>(nameof(DialogState)), cancellationToken);
-
-        //             //await turnContext.SendActivityAsync(MessageFactory.Text($"Hello world!"), cancellationToken);
-        //         }
-        //     }
-        // }
-
-        protected override async Task OnMessageActivityAsync(ITurnContext<IMessageActivity> turnContext, CancellationToken cancellationToken)
-        {
-
-            // Run the Dialog with the new message Activity.
-            await Dialog.RunAsync(turnContext, ConversationState.CreateProperty<DialogState>(nameof(DialogState)), cancellationToken);
-        }
-
-        
-    }
-}
