+++ FoodFite/Program.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 
-namespace EmptyBot
+namespace FoodFite
 {
      public class Program
     {
